Turn off adjustment by hotness
In the current implementation, we use hotness to decide victime page node in
the two level cache. The hotness is the sum of last-access timestamps of entry
nodes. This approach seems OK at first glance. But it is actually not OK. This
is what I see in my experiment:

For btrfs, f2fs and ext4, almost all the time, there are 4 page nodes in the
cache (sometimes 3), pages nodes m_vpn 0 1 2 are the hotest because they have
many entry nodes in them (recall that hotness is a sum). The fourth page node
has the latest accessed entry node, but it is the coldest page node because the
page node only has one entry in it. Because TPFTL chooses the coldest page
nodes's coldest entry to replace, the fourth page node's entry (the only entry)
is always chosen as the victim, although it is just accessed.

The behavior above shows the **fundamental problem of TPFTL two-level cache:
the coldest `page_node` may not have the coldest entries. In contrast, it may
have the most recently accessed entry**. I guess TPFTL's motivation of using
page node hotness to choose victim is to find coldest page node so it is good
for batch write back. But such a design have unexpected behavior described
above.

Using number of hits would lead to the same problem.