Fix the global gc race full merge problem.
The problem cause is in the recycle_empty_data_block(). It tries to recycle newly
allocated data block, which is actively being written to. There are two case when
this happens. First, when the block is brand new (all pages are erased). Second,
when the first k pages are invalidated and the rest is erased.