Fix key error when recycle empty data block
The reason for the key error is because a block is in used data block pool
but not in data block mapping. Similar thing can happen to log blocks. A block
can be in log used block pool but not in log block mapping. This can happen because,
well, there can be a moment before a block is popped out from free block pool and
not in any mapping yet. The solutio of this is, when you need to find a TRUE
data block, you need to make sure 1. the data block is in used data block pool, 2. the
block is in data block mapping.